[共通] fix warnings

+++ Siv3D/src/Siv3D/Font/FontFace.cpp
@@ -170,7 +170,7 @@ namespace s3d
 		{
 			m_property.familiyName		= Unicode::Widen(m_face->family_name);
 			m_property.styleName		= Unicode::Widen(m_face->style_name);
-			m_property.numGlyphs		= m_face->num_glyphs;
+			m_property.numGlyphs		= static_cast<uint32>(m_face->num_glyphs);
 			m_property.hasColor			= FT_HAS_COLOR(m_face);
 			m_property.fontPixelSize	= pixelSize;
 			m_property.style			= style;
